// web/src/openapi/config/zfcp.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * zFCP config.
 */
export interface Config {
  /**
   * List of zFCP controllers.
   */
  controllers?: string[];
  /**
   * List of zFCP devices.
   */
  devices?: Device[];
}
export interface Device {
  /**
   * zFCP controller channel id.
   */
  channel: string;
  /**
   * WWPN of the target port.
   */
  wwpn: string;
  /**
   * LUN of the SCSI device.
   */
  lun: string;
  /**
   * Whether to activate the device.
   */
  active?: boolean;
}